+++ packages/lumx-core/src/js/components/Skeleton/index.ts
@@ -1,0 +1,24 @@
+export {
+    SkeletonCircle,
+    type SkeletonCircleProps,
+    COMPONENT_NAME as SKELETON_CIRCLE_COMPONENT_NAME,
+    CLASSNAME as SKELETON_CIRCLE_CLASSNAME,
+    DEFAULT_PROPS as SKELETON_CIRCLE_DEFAULT_PROPS,
+} from './SkeletonCircle';
+
+export {
+    SkeletonRectangle,
+    type SkeletonRectangleProps,
+    SkeletonRectangleVariant,
+    COMPONENT_NAME as SKELETON_RECTANGLE_COMPONENT_NAME,
+    CLASSNAME as SKELETON_RECTANGLE_CLASSNAME,
+    DEFAULT_PROPS as SKELETON_RECTANGLE_DEFAULT_PROPS,
+} from './SkeletonRectangle';
+
+export {
+    SkeletonTypography,
+    type SkeletonTypographyProps,
+    COMPONENT_NAME as SKELETON_TYPOGRAPHY_COMPONENT_NAME,
+    CLASSNAME as SKELETON_TYPOGRAPHY_CLASSNAME,
+    DEFAULT_PROPS as SKELETON_TYPOGRAPHY_DEFAULT_PROPS,
+} from './SkeletonTypography';