+++ packages/lumx-vue/src/components/skeleton/SkeletonCircle.test.ts
@@ -1,0 +1,11 @@
+import { render } from '@testing-library/vue';
+import BaseSkeletonCircleTests from '@lumx/core/js/components/Skeleton/SkeletonCircleTests';
+import SkeletonCircle from './SkeletonCircle';
+import type { SkeletonCircleProps } from '@lumx/core/js/components/Skeleton';
+
+describe('<SkeletonCircle>', () => {
+    // Run core tests
+    BaseSkeletonCircleTests({
+        render: (props: SkeletonCircleProps) => render(SkeletonCircle, { props }),
+    });
+});